@@ -4,11 +4,11 @@
   "private": true,
   "dependencies": {
     "prop-types": "^15.7.2",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-redux": "^5.1.1",
-    "react-scripts": "1.0.10",
-    "redux": "^3.7.2"
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4",
+    "react-redux": "^7.1.3",
+    "react-scripts": "2.1.8",
+    "redux": "^4.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -16,5 +16,11 @@
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
-  "devDependencies": {}
+  "devDependencies": {},
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
