@@ -1,7 +1,7 @@
 import React from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { actionCreators } from '../features/redux/actions'
+import { add, remove, toggle, update } from '../features/redux/actions'
 
 import List from '../features/components/List'
 import Input from '../features/components/Input'
@@ -15,18 +15,22 @@ const styles = {
 };
 
 function App (props) {
-    const { todoList, add, remove, toggle } = props;
+    const { todoList, add, remove, toggle, update } = props;
 
-    const onAddTodo = (text) => {
-        add(text)
+    const onAddTodo = (todo) => {
+        add(todo)
     };
 
-    const onRemoveTodo = (index) => {
-        remove(index)
+    const onRemoveTodo = (id) => {
+        remove(id)
     };
 
-    const onToggleTodo = (index) => {
-        toggle(index)
+    const onToggleTodo = (id) => {
+        toggle(id)
+    };
+
+    const onUpdateItem = (id, todo) => {
+        update(id, todo)
     };
 
     return (
@@ -42,6 +46,7 @@ function App (props) {
                 list         = { todoList }
                 onRemoveItem = { onRemoveTodo }
                 onToggleItem = { onToggleTodo }
+                onUpdateItem = { onUpdateItem}
             />
         </div>
     )
@@ -49,12 +54,12 @@ function App (props) {
 
 function mapStateToProps(state) {
     return {
-        todoList: state.todos
+        todoList: state
     }
 }
 
 function mapDispatchToProps(dispatch) {
-    return bindActionCreators(actionCreators, dispatch)
+    return bindActionCreators({ add, remove, toggle, update }, dispatch)
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(App);
