@@ -1,30 +1,30 @@
 import types from '../constants'
 
 // Helper functions to dispatch actions, optionally with payloads
-export const actionCreators = {
-    add: todo => {
-        return {
-            type: types.ADD,
-            payload: todo
-        };
-    },
-    remove: todo => {
-        return {
-            type: types.REMOVE,
-            payload: todo
-        };
-    },
-    toggle: todo => {
-        return {
-            type: types.TOGGLE,
-            payload: todo
-        };
-    },
-    save: (id, text)=> {
-        return {
-            type: types.SAVE,
-            payload: {id, text}
-        }
-    }
+export const add = (todo) => {
+    return {
+        type: types.ADD,
+        payload: todo
+    };
+};
 
+export const remove = (todo) => {
+    return {
+        type: types.REMOVE,
+        payload: todo
+    };
+};
+
+export const toggle = (todo) => {
+    return {
+        type: types.TOGGLE,
+        payload: todo
+    };
+};
+
+export const update = (id, text)=> {
+    return {
+        type: types.UPDATE,
+        payload: { id, text }
+    }
 };
